requirements.txt
datasets
langchain
langchain_community
langchain-openai
langchain-tavily
langgraph
wikipedia
openpyxl
openai-whisper